HTTPS Client example without peer verification.

+++ source/corvusoft/restbed/uri.cpp
@@ -182,13 +182,13 @@ namespace restbed
     {
         return encode( Bytes( value.begin( ), value.end( ) ) );
     }
-
+    
     string Uri::encode_parameter( const string& value )
     {
         return encode( value );
     }
     
-    int Uri::get_port( void ) const
+    uint16_t Uri::get_port( void ) const
     {
         smatch match;
         string port = String::empty;
